Removed invalid S2SB pin and update README.md.
Removed the URL to EvanVS's S2SB server, which I initially believed to be invalid. Also removed the two S2SB Python files as they were miscellaneous scripts for something else. However, the cap2rss.py script is not functioning correctly with the apps.fema.gov feed. I will not be putting in changes to fix this. View the README.